+++ CHANGELOG.md
@@ -16,6 +16,10 @@
 
 - BREAKING: Stop supporting deprecated 'HiPhish/nvim-ts-rainbow2'.
 
+## mini.misc
+
+- FEATURE: `setup_auto_root()` and `find_root()` now have `fallback` argument to be applied when no root is found with `vim.fn.find()`.
+
 ## mini.pick
 
 Introduction of a new module.